change: use u32 to represent node id's instead of u8
With u8's, up to 256 nodes were possible. Warnet might want to use more
nodes. Using an u32 instead of an u8 significatly increases the
theoretical maximum.

+++ src/node.rs
@@ -230,7 +230,7 @@ pub trait Node: Sync {
 
 #[derive(Hash, Clone)]
 pub struct NodeInfo {
-    pub id: u8,
+    pub id: u32,
     pub name: String,
     pub description: String,
 }
